moved log_sp_error to nta_aux module
